+++ spec/features/projects/users_can_view_a_list_of_assigned_projects_spec.rb
@@ -1,6 +1,6 @@
 require "rails_helper"
 
-RSpec.feature "Viewing assigend projects" do
+RSpec.feature "Viewing assigned projects" do
   before do
     sign_in_with_user(user)
     mock_successful_api_response_to_create_any_project
@@ -15,8 +15,6 @@ RSpec.feature "Viewing assigend projects" do
   let!(:completed_unassigned_project) { create(:conversion_project, urn: 121583, assigned_to: nil, completed_at: Date.yesterday) }
   let!(:in_progress_unassigned_project) { create(:conversion_project, urn: 115652, assigned_to: nil) }
 
-  let!(:completed_assigned_project) { create(:conversion_project, urn: 114067, assigned_to: user, completed_at: Date.yesterday) }
-
   context "when there are no projects" do
     let(:user) { create(:user, :caseworker) }
 
@@ -24,17 +22,26 @@ RSpec.feature "Viewing assigend projects" do
       visit user_in_progress_projects_path
 
       expect(page).to have_content(I18n.t("project.index.empty"))
+
+      visit user_completed_projects_path
+
+      expect(page).to have_content(I18n.t("project.table.completed.empty"))
     end
   end
 
   context "when there are projects" do
     let!(:in_progress_assigned_project) { create(:conversion_project, urn: 103835, assigned_to: user) }
+    let!(:completed_assigned_project) { create(:conversion_project, urn: 114067, assigned_to: user, completed_at: Date.yesterday) }
 
     context "when signed in as a Regional caseworker" do
       let(:user) { create(:user, :caseworker) }
 
       scenario "they can view all in progress projects" do
         view_in_progress_projects
+      end
+
+      scenario "they can view assigned completed" do
+        view_completed_projects
       end
     end
 
@@ -44,6 +51,10 @@ RSpec.feature "Viewing assigend projects" do
       scenario "they can view all in progress projects" do
         view_in_progress_projects
       end
+
+      scenario "they can view assigned completed" do
+        view_completed_projects
+      end
     end
 
     context "when signed in as a Regional delivery officer" do
@@ -51,6 +62,10 @@ RSpec.feature "Viewing assigend projects" do
 
       scenario "they can view all in progress projects" do
         view_in_progress_projects
+      end
+
+      scenario "they can view assigned completed" do
+        view_completed_projects
       end
     end
 
@@ -68,8 +83,23 @@ RSpec.feature "Viewing assigend projects" do
 
         expect(page).not_to have_content(completed_other_user_project.urn)
         expect(page).not_to have_content(completed_unassigned_project.urn)
+      end
+    end
 
-        expect(page).not_to have_content(completed_assigned_project.urn)
+    def view_completed_projects
+      visit user_completed_projects_path
+
+      expect(page).to have_content(I18n.t("project.user.completed.title"))
+
+      within("tbody") do
+        expect(page).to have_content(completed_assigned_project.urn)
+        expect(page).not_to have_content(in_progress_assigned_project.urn)
+
+        expect(page).not_to have_content(in_progress_other_user_project.urn)
+        expect(page).not_to have_content(in_progress_unassigned_project.urn)
+
+        expect(page).not_to have_content(completed_other_user_project.urn)
+        expect(page).not_to have_content(completed_unassigned_project.urn)
       end
     end
   end